requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.0.0
requests>=2.31.0
